dash-data-app-obo-user/requirements.txt
dash
dash-bootstrap-components
pandas
plotly
databricks-sql-connector
databricks-sdk
python-dotenv
dash-ag-grid
databricks-sdk[databricks-sql]
pyyaml